+++ lib/core-server/src/core-presets.test.ts
@@ -1,51 +1,44 @@
 import 'jest-specific-snapshot';
 import path from 'path';
-import { Configuration } from 'webpack';
+import webpack, { Configuration } from 'webpack';
 import Cache from 'file-system-cache';
-import { resolvePathInStorybookCache } from '@storybook/core-common';
-import { executor as previewExecutor } from '@storybook/builder-webpack4';
-import { executor as managerExecutor } from './manager/builder';
 
 import { buildDevStandalone } from './build-dev';
 import { buildStaticStandalone } from './build-static';
+import { resolvePathInStorybookCache } from './utils/resolve-path-in-sb-cache';
+import reactOptions from '../../../../app/react/src/server/options';
 
-import reactOptions from '../../../app/react/src/server/options';
+const TIMEOUT = 10000;
 
-jest.mock('@storybook/builder-webpack5', () => {
-  const actualBuilder = jest.requireActual('@storybook/builder-webpack5');
-  // MUTATION! we couldn't mock webpack5, so we added a level of indirection instead
-  actualBuilder.executor.get = jest.fn();
-  return actualBuilder;
+const mockStats = {
+  hasErrors: () => false,
+  hasWarnings: () => false,
+  toJson: () => ({ warnings: [], errors: [] }),
+};
+const webpackStub = (cb) => {
+  return cb(null, mockStats);
+};
+const middlewareStub = (_req, _resp, next) => next();
+middlewareStub.waitUntilValid = (cb) => cb(mockStats);
+jest.mock('webpack-hot-middleware', () => () => middlewareStub);
+jest.mock('webpack-dev-middleware', () => () => middlewareStub);
+jest.mock('webpack', () => {
+  const actualWebpack = jest.requireActual('webpack');
+  return {
+    __esModule: true, // this property makes it work
+    default: jest.fn((...args) => {
+      const compiler = actualWebpack(...args);
+      compiler.watch = jest.fn(webpackStub);
+      compiler.run = jest.fn(webpackStub);
+      return compiler;
+    }),
+    ...actualWebpack,
+  };
 });
 
-jest.mock('@storybook/builder-webpack4', () => {
-  const actualBuilder = jest.requireActual('@storybook/builder-webpack4');
-  // MUTATION! we couldn't mock webpack5, so we added a level of indirection instead
-  actualBuilder.executor.get = jest.fn();
-  return actualBuilder;
-});
-
-jest.mock('./manager/builder', () => {
-  const actualBuilder = jest.requireActual('./manager/builder');
-  // MUTATION!
-  actualBuilder.executor.get = jest.fn();
-  return actualBuilder;
-});
-
-jest.mock('cpy', () => () => Promise.resolve());
 jest.mock('http', () => ({
   ...jest.requireActual('http'),
   createServer: () => ({ listen: (_options, cb) => cb() }),
-}));
-jest.mock('@storybook/node-logger', () => ({
-  logger: {
-    info: jest.fn(),
-    warn: jest.fn(),
-    error: jest.fn(),
-  },
-}));
-jest.mock('./utils/output-startup-information', () => ({
-  outputStartupInformation: jest.fn(),
 }));
 
 const cache = Cache({
@@ -54,14 +47,13 @@ const cache = Cache({
 });
 
 const managerOnly = false;
-const options = {
+const baseOptions = {
   ...reactOptions,
   ignorePreview: managerOnly,
   // FIXME: this should just be ignorePreview everywhere
   managerOnly, // production
   docsMode: false,
   cache,
-  configDir: path.resolve(`${__dirname}/../../../examples/react-ts`),
   outputDir: `${__dirname}/storybook-static`, // production
   ci: true,
   managerCache: false,
@@ -78,9 +70,6 @@ const cleanRoots = (obj): any => {
   if (typeof obj === 'object') {
     return Object.fromEntries(
       Object.entries(obj).map(([key, val]) => {
-        if (key === 'version' && typeof val === 'string') {
-          return [key, '*'];
-        }
         return [key, cleanRoots(val)];
       })
     );
@@ -91,47 +80,82 @@ const cleanRoots = (obj): any => {
 const prepareSnap = (fn: any, name): Pick<Configuration, 'module' | 'entry' | 'plugins'> => {
   const call = fn.mock.calls.find((c) => c[0].name === name);
   if (!call) return null;
+
+  const keys = Object.keys(call[0]);
   const { module, entry, plugins } = call[0];
 
-  return cleanRoots({ module, entry, plugins: plugins.map((p) => p.constructor.name) });
+  return cleanRoots({ keys, module, entry, plugins: plugins.map((p) => p.constructor.name) });
 };
 
 const snap = (name: string) => `__snapshots__/${name}`;
 
-describe('manager', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-    cache.clear();
-  });
-  it('dev mode', async () => {
-    await buildDevStandalone({ ...options, ignorePreview: true });
+describe.each([
+  ['cra-ts-essentials'],
+  ['vue-3-cli'],
+  ['angular-cli'],
+  ['web-components-kitchen-sink'],
+  ['html-kitchen-sink'],
+])('%s', (example) => {
+  const options = {
+    ...baseOptions,
+    configDir: path.resolve(`${__dirname}/../../../../examples/${example}/.storybook`),
+  };
 
-    const managerConfig = prepareSnap(managerExecutor.get, 'manager');
-    expect(managerConfig).toMatchSpecificSnapshot(snap('manager-dev'));
-  });
-  it('production mode', async () => {
-    await buildStaticStandalone({ ...options, ignorePreview: true });
+  describe('manager', () => {
+    beforeEach(() => {
+      jest.clearAllMocks();
+      cache.clear();
+    });
+    it(
+      'dev mode',
+      async () => {
+        const result = await buildDevStandalone(options);
+        expect(webpack).toHaveBeenCalled();
 
-    const managerConfig = prepareSnap(managerExecutor.get, 'manager');
-    expect(managerConfig).toMatchSpecificSnapshot(snap('manager-prod'));
-  });
-});
+        const managerConfig = prepareSnap(webpack, 'manager');
+        expect(managerConfig).toMatchSpecificSnapshot(snap(`${example}_manager-dev`));
+      },
+      TIMEOUT
+    );
+    it(
+      'production mode',
+      async () => {
+        const result = await buildStaticStandalone(options);
+        expect(webpack).toHaveBeenCalled();
 
-describe('preview', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-    cache.clear();
+        const managerConfig = prepareSnap(webpack, 'manager');
+        expect(managerConfig).toMatchSpecificSnapshot(snap(`${example}_manager-prod`));
+      },
+      TIMEOUT
+    );
   });
-  it('dev mode', async () => {
-    await buildDevStandalone({ ...options, managerCache: true });
 
-    const previewConfig = prepareSnap(previewExecutor.get, 'preview');
-    expect(previewConfig).toMatchSpecificSnapshot(snap('preview-dev'));
-  });
-  it('production mode', async () => {
-    await buildStaticStandalone({ ...options, managerCache: true });
+  describe('preview', () => {
+    beforeEach(() => {
+      jest.clearAllMocks();
+      cache.clear();
+    });
+    it(
+      'dev mode',
+      async () => {
+        const result = await buildDevStandalone(options);
+        expect(webpack).toHaveBeenCalled();
 
-    const previewConfig = prepareSnap(previewExecutor.get, 'preview');
-    expect(previewConfig).toMatchSpecificSnapshot(snap('preview-prod'));
+        const previewConfig = prepareSnap(webpack, 'preview');
+        expect(previewConfig).toMatchSpecificSnapshot(snap(`${example}_preview-dev`));
+      },
+      TIMEOUT
+    );
+    it(
+      'production mode',
+      async () => {
+        const result = await buildStaticStandalone(options);
+        expect(webpack).toHaveBeenCalled();
+
+        const previewConfig = prepareSnap(webpack, 'preview');
+        expect(previewConfig).toMatchSpecificSnapshot(snap(`${example}_preview-prod`));
+      },
+      TIMEOUT
+    );
   });
 });