+++ lib/ui/src/components/preview/utils/stringifyQueryParams.tsx
@@ -1,4 +1,4 @@
 import qs from 'qs';
 
 export const stringifyQueryParams = (queryParams: Record<string, string>) =>
-  qs.stringify(queryParams, { addQueryPrefix: true, encode: true }).replace(/^\?/, '&');
+  qs.stringify(queryParams, { addQueryPrefix: true, encode: false }).replace(/^\?/, '&');